content/media/test/test_encryptedMediaExtensions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var manager = new MediaTestManager;

const KEYSYSTEM_TYPE = "org.w3.clearkey";

function bail(message)
{
  return function(err) {
    ok(false, message);
    if (err) {
      info(err);
    }
    SimpleTest.finish();
  }
}

function ArrayBufferToString(arr)
{
  var str = '';
  var view = Uint8Array(arr);
  for (var i = 0; i < view.length; i++) {
    str += String.fromCharCode(view[i]);
  }
  return str;
}

function StringToArrayBuffer(str)
{
  var arr = ArrayBuffer(str.length);
  var view = Uint8Array(arr);
  for (var i = 0; i < str.length; i++) {
    view[i] = str.charCodeAt(i);
  }
  return arr;
}

function Base64ToHex(str)
{
  var bin = window.atob(str.replace(/-/g, "+").replace(/_/g, "/"));
  var res = "";
  for (var i = 0; i < bin.length; i++) {
    res += ("0" + bin.charCodeAt(i).toString(16)).substr(-2);
  }
  return res;
}

function HexToBase64(hex)
{
  var bin = "";
  for (var i = 0; i < hex.length; i += 2) {
    bin += String.fromCharCode(parseInt(hex.substr(i, 2), 16));
  }
  return window.btoa(bin).replace(/=/g, "").replace(/\+/g, "-").replace(/\//g, "_");
}

function UpdateSessionFunc(test) {
  return function(ev) {
    var msgStr = ArrayBufferToString(ev.message);
    var msg = JSON.parse(msgStr);

    info("got message from CDM: " + msgStr);
    is(msg.type, test.sessionType, "Key session type should match");
    ok(msg.kids, "message event should contain key ID array");

    var outKeys = [];

    for (var i = 0; i < msg.kids.length; i++) {
      var id64 = msg.kids[i];
      var idHex = Base64ToHex(msg.kids[i]).toLowerCase();
      var key = test.keys[idHex];

      if (key) {
        info("found key " + key + " for key id " + idHex);
        outKeys.push({
          "kty":"oct",
          "alg":"A128KW",
          "kid":id64,
          "k":HexToBase64(key)
        });
      } else {
        bail("Couldn't find key for key id " + idHex);
      }
    }

    var update = JSON.stringify({
      "keys" : outKeys,
      "type" : msg.type
    });
    info("sending update message to CDM: " + update);

    ev.target.update(StringToArrayBuffer(update)).then(function() {
      info("MediaKeySession update ok!");
    }, bail("MediaKeySession update failed"));
  }
}

function PlayFragmented(test, elem)
{
  var ms = new MediaSource();
  elem.src = URL.createObjectURL(ms);

  var sb;
  var curFragment = 0;

  function addNextFragment() {
    if (curFragment >= test.fragments.length) {
      ms.endOfStream();
      elem.play();
      return;
    }

    var fragmentFile = test.fragments[curFragment++];

    var req = new XMLHttpRequest();
    req.open("GET", fragmentFile);
    req.responseType = "arraybuffer";

    req.addEventListener("load", function() {
      sb.appendBuffer(new Uint8Array(req.response));
    });

    info("fetching resource " + fragmentFile);
    req.send(null);
  }

  ms.addEventListener("sourceopen", function () {
    sb = ms.addSourceBuffer(test.type);
    sb.addEventListener("updateend", addNextFragment);

    addNextFragment();
  });
}

function PlayTest(test, elem)
{
  if (test.fragments) {
    PlayFragmented(test, elem);
    return;
  }

  // This file isn't fragmented; set the media source normally.
  elem.src = test.name;
  elem.play();
}

function KeysChangeFunc(session, keys) {
  session.keyIdsReceived = [];
  for (var keyid in keys) {
    info("Set " + keyid + " to false in session.keyIdsReceived");
    session.keyIdsReceived[keyid] = false;
  }
  return function(ev) {
    var session = ev.target;
    session.gotKeysChanged = true;
    session.getUsableKeyIds().then(function(keyIds) {
      for (var k = 0; k < keyIds.length; k++) {
        var kid = Base64ToHex(window.btoa(ArrayBufferToString(keyIds[k])));
        ok(kid in session.keyIdsReceived, "session.keyIdsReceived contained " + kid + " as expected.");
        session.keyIdsReceived[kid] = true;
      }
    }, bail("Failed to get keyIds"));
  }
}

function startTest(test, token)
{
  manager.started(test._token);

  var v = document.createElement("video");
  var gotEncrypted = false;
  var gotPlaying = false;

  v.addEventListener("encrypted", function(ev) {
    gotEncrypted = true;

    info(token + " got encrypted event");
    ok(MediaKeys.isTypeSupported(KEYSYSTEM_TYPE, ev.initDataType, test.type),
       token + " MediaKeys should support this keysystem");

    MediaKeys.create(KEYSYSTEM_TYPE).then(function(mediaKeys) {
      info(token + " created MediaKeys object ok");
      mediaKeys.sessions = [];
      return v.setMediaKeys(mediaKeys);
    }, bail("failed to create MediaKeys object")).then(function() {
      info(token + " set MediaKeys on <video> element ok");

      ok(MediaKeys.isTypeSupported(KEYSYSTEM_TYPE, ev.initDataType, test.type),
         "MediaKeys should still support keysystem after CDM created...");

      var session = v.mediaKeys.createSession(test.sessionType);
      v.mediaKeys.sessions.push(session);
      session.addEventListener("keyschange", KeysChangeFunc(session, test.keys), false);
      session.addEventListener("message", UpdateSessionFunc(test));
      session.generateRequest(ev.initDataType, ev.initData).then(function() {
      }, bail(token + " Failed to initialise MediaKeySession"));

    }, bail(token + " Failed to set MediaKeys on <video> element"));
  });

  v.addEventListener("playing", function () { gotPlaying = true; });

  v.addEventListener("ended", function(ev) {
    ok(true, token + " got ended event");
    manager.finished(test._token);

    ok(gotEncrypted, token + " encrypted event should have fired");
    ok(gotPlaying, token + " playing event should have fired");

    ok(Math.abs(test.duration - v.duration) < 0.1,
       token + " Duration of video should be corrrect");
    ok(Math.abs(test.duration - v.currentTime) < 0.1,
       token + " Current time should be same as duration");
    // Verify all sessions had all keys went sent the to the CDM usable, and thus
    // that we received keyschange event(s).
    var sessions = v.mediaKeys.sessions;
    is(sessions.length, 1, "should have 1 session");
    for (var i = 0; i < sessions.length; i++) {
      var session = sessions[i];
      ok(session.gotKeysChanged, "Should have received at least one keychange event");
      for (var kid in session.keyIdsReceived) {
        ok(session.keyIdsReceived[kid], "key with id " + kid + " was usable as expected");
      }
    }
  });

  v.addEventListener("error", bail(token + " got error event"));

  PlayTest(test, v);
}

function testIsTypeSupported()
{
  var t = MediaKeys.isTypeSupported;
  const clearkey = "org.w3.clearkey";
  ok(!t("bogus", "bogon", "video/bogus"), "Invalid type.");
  ok(t(clearkey), "ClearKey supported.");
  ok(!t(clearkey, "bogus"), "ClearKey bogus initDataType not supported.");
  ok(t(clearkey, "cenc"), "ClearKey/cenc should be supported.");
  ok(!t(clearkey, "cenc", "bogus"), "ClearKey/cenc bogus content type should be supported.");
  ok(t(clearkey, "cenc", 'video/mp4'), "ClearKey/cenc video/mp4 supported.");
  ok(t(clearkey, "cenc", 'video/mp4; codecs="avc1.4d4015,mp4a.40.2"'), "ClearKey/cenc H.264/AAC supported.");
  ok(t(clearkey, "cenc", 'audio/mp4'), "ClearKey/cenc audio/mp4 supported.");
  ok(t(clearkey, "cenc", 'audio/mp4; codecs="mp4a.40.2"'), "ClearKey/cenc AAC LC supported.");
}

function beginTest() {
  testIsTypeSupported();
  manager.runTests(gEMETests, startTest);
}

var prefs = [
  [ "media.mediasource.enabled", true ],
  [ "media.mediasource.ignore_codecs", true ],
];

if (/Linux/.test(navigator.userAgent) ||
    !document.createElement('video').canPlayType("video/mp4")) {
  // XXX remove once we have mp4 PlatformDecoderModules on all platforms.
  prefs.push([ "media.fragmented-mp4.exposed", true ]);
  prefs.push([ "media.fragmented-mp4.use-blank-decoder", true ]);
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({ "set" : prefs }, beginTest);
</script>
</pre>
</body>
</html>
